requirements.txt
pytest
requests
playwright
pytest-playwright
allure-pytest
python-dotenv
pytest-xdist
pytest-html
pyyaml
pytest-ordering
autopep8
